backend/requirements.txt
fastapi
uvicorn
websockets
pydantic
google-generativeai
pytest
httpx
